add classes for the jina models
