+++ src/__init__.py
@@ -1,4 +1,5 @@
 from .wikidata_dumpreader import WikidataDumpReader
 from .wikidataDB import WikidataEntity, WikidataID
-from .wikidataEmbed import WikidataTextifier, JinaAIEmbedder, JinaAIReranker
+from .wikidataEmbed import WikidataTextifier
+from .JinaAI import JinaAIEmbedder, JinaAIReranker
 from .wikidataRetriever import AstraDBConnect
